add aclrt

+++ src/llm_datadist/common/rank_table_generator_v2.cc
@@ -106,8 +106,8 @@ ge::Status RankTableGeneratorV2::MergeRankTable(int32_t local_device_id,
     LLM_CHK_BOOL_RET_STATUS(server.device_list.size() == 1U, ge::LLM_PARAM_INVALID,
                            "Please check local option:%s, it only supports one device that is used.",
                            llm_datadist::OPTION_LOCAL_COMM_RES);
-    uint32_t phy_device_id = 0U;
-    LLM_CHK_RT_RET(aclrtGetPhyDevIdByLogicDevId(static_cast<uint32_t>(local_device_id), &phy_device_id));
+    int32_t phy_device_id = 0U;
+    LLM_CHK_RT_RET(aclrtGetPhyDevIdByLogicDevId(local_device_id, &phy_device_id));
     LLM_CHK_BOOL_RET_STATUS(std::to_string(phy_device_id) == server.device_list[0].device_id,
                            ge::LLM_PARAM_INVALID,
                            "Please check local option:%s, device_id:%s should be %u, logic device id:%d.",
@@ -220,8 +220,8 @@ ge::Status RankTableGeneratorV2::GenerateLocalCommRes(const std::string &server_
   rank_table_v2::ServerInfo server_info{};
   server_info.server_id = server_id;
   rank_table_v2::DeviceInfo device_info{};
-  uint32_t phy_device_id = 0U;
-  LLM_CHK_RT_RET(aclrtGetPhyDevIdByLogicDevId(static_cast<uint32_t>(device_id), &phy_device_id));
+  int32_t phy_device_id = 0U;
+  LLM_CHK_RT_RET(aclrtGetPhyDevIdByLogicDevId(device_id, &phy_device_id));
   device_info.device_id = std::to_string(phy_device_id);
   LLM_CHK_STATUS_RET(LocalCommResGenerator::GetDeviceIp(phy_device_id, device_info.device_ip),
                     "Failed to get device_ip, phy_device_id:%u", phy_device_id);
